[feat] Add application server logic
Splits the current directory structure into three parts:

1. `cmd` for the public-facing API.
2. `internal/pkg/http` for the application server.
3. `internal/cmd/api` for the internal server struct that is used by
   `cmd/api`.

The `server` struct exposes a visible method, `StartApplicationServer`
that can be used outside the `api` package. It runs a shutdown method in
a goroutine that waits to receive a `SIGINT` or `SIGTERM` signal. It
relies on `signal.Notify` to circumvent the *default* behavior for
synchronous signals in Go. Without this method call, the signal would be
converted into a runtime panic which would be caught by the main thread,
and the HTTP server would not be shutdown gracefully.

The `cmd/api/main.go` starts the application server and blocks on that
call. In the future, if we add an admin server, we should refactor this
functionality to start up everything relevant to running the server.
