+++ cmd/api/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"card-match/internal/cmd/api"
+)
+
+func main() {
+	server := api.NewServer()
+	server.StartApplicationServer()
+}
